Use Conan packages for libnest2d, pynest2d, savitar, arcus and curaengine
It turned out to be troublesome to have a hybrid "legacy" conan pipeline
build environment. In the regards of how compiler flags and cmake
variables were shared so I had to create conan packages and their
corresponding GitHub workflows for the C++ packages in our care.

These workflow are still rudementary but should server our needs for
now.

Contributes to CURA-9365